+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT RELEASE
 
 - Fixes a bug where `false` booleans in the GitHub Actions configuration would still pass through truthy. This is because GitHub Actions passes all env vars to the action as a string. We simply check for false strings and coerce them to False bools now (closes #16)
+- Fixes a bug that would kill the action when no `license` or `description` was set on a repo (closes #17)
 
 ## v0.12.0 (2022-08-12)